pyproject.toml
[project]
name = "demo-langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]
